Restart dead connections:
when a connection is detected down, we delete it. But until now we have not attempted to re-create it if the interface/ route comes up again. This commit implements some additional info and methods which makes it possible to detect if a connection has been deleted, so we can re-create it.