Packet sorter: Better handling of sequence holes.
In cases where the packet sorter queue contains multiple non-contiguous sequence holes, packets could end up waiting n*deadline where n is the number of sequence holes.

We now save packets in the sorter with a timestamp, so we can keep track of how log a packet has been waiting. If the wait time exceeds the deadline we trigger the deadline timer instantly.