Bugfix: Connection timeout setting not used
When a user set the connection timeout in the configuration file, we correctly read the setting and propagated it. However, it was only used on the first timeout event. After that we used a hardcoded 10 second interval.

This has now been corrected, such that the user configuration is used every time.